// src/interface/interface.tsx
export interface IRequestAccess {
    _id: string;
    rojifiId: string;
    firstname: string;
    lastname: string;
    middlename: string;
    email: string;
    businessName: string;
    businessWebsite: string;
    message: string;
    weeklyVolume: number;
    phoneCode: string;
    phoneNumber: string;
    address: string;
    city: string;
    state: string;
    country: string;
    postalCode: string;
    agreement: boolean;
    approved: boolean;
    deleted: boolean;
    archived: boolean;
    approvedAt: Date | null;
    deletedAt: Date | null;
    archivedAt: Date | null;
    metadata: Record<string, any>;
}

export interface IResponse<Data = any, Error = any> {
    code: number;
    status: Status;
    message: string;
    data?: Data;
    error?: Error;
    pagination?: IPagination;
    timestamp: string;
    requestId: string;
    copyright: ICopyright;
    help: Array<string>;
    docs: string;
    version: string;
    handshake?: string;
}

export interface IPagination {
    page: number;
    limit: number;
    total: number;
    totalPages: number;
}

export interface ICopyright {
    year: string;
    holder: string;
    license: string;
    licenseUrl: string;
}
// ...
export interface IUser extends IRequestAccess {
    post_no_debit: string;
    address_line_one: string;
    address_line_two: string;
    username: string;
    fullName: string;
    isEmailVerified: boolean;
    key: string;
    phoneNumberHash: string;
    isPhoneNumberVerified: boolean;
    dateOfBirth: string;
    pin: string;
    mnemonic: string;
    referralCode: string;
    password: string;
    //////////////////////////////////

    cacCertOfIncoporation: string,
    memorandumArticlesOfAssociation: string,
    cacStatusReport: string,
    proofOfAddress: string,

    cacCertOfIncoporationIsVerified: boolean,
    memorandumArticlesOfAssociationIsVerified: boolean,
    cacStatusReportIsVerified: boolean,
    proofOfAddressIsVerified: boolean,

    cacCertOfIncoporationVerifiedAt: Date | null,
    memorandumArticlesOfAssociationVerifiedAt: Date | null,
    cacStatusReportVerifiedAt: Date | null,
    proofOfAddressVerifiedAt: Date | null,

    requested: {
        otcdesk: RequestStatus;
        virtualCard: RequestStatus;
    }

    //////////////////////////////////
    biometricType: BiometricType;
    biometric: string;
    biometricVerified: boolean;
    biometricVerifiedAt: Date;
    biometricEnabled: boolean;
    isVerificationComplete: boolean;
    loginCount: number;
    loginLastAt: Date;
    loginLastIp: string;
    loginLastDevice: string;
    twoFactorSecret: string;
    twoFactorURL: string;
    twoFactorEnabled: boolean;
    twoFactorVerified: boolean;
    twoFactorVerifiedAt: Date;
    isSuspecious: boolean;
    isActive: boolean;
    isSuspended: boolean;
    isBanned: boolean;
    userType: UserType;
    refreshToken: string;
    deviceToken: string;
    createdAt: Date;
    updatedAt: Date;
    passkey: string;
    passkeyEnabled: boolean;
    passkeyVerified: boolean;
    passkeyVerifiedAt: Date;
    tier: AccountTier;
    firstDepositConfirmed: boolean;
    deletedBy: string | null;
    comparePassword(password: string): Promise<boolean>;
    comparePin(pin: string): Promise<boolean>;
    comparePasskey(passkey: string): Promise<boolean>;
    compareBiometric(biometric: string): Promise<boolean>;
    compareDeviceToken(deviceToken: string): Promise<boolean>;
}